fix: 补充vlw中当前插件用到的api，后续抽空补齐其他api

+++ engine/robot/event.go
@@ -17,5 +17,5 @@ type Message struct {
 	FromGroupName string
 	Msg           string
 	MsgId         string
-	MsgType       int
+	MsgType       int64
 }
